TOOL-8 Reorganize InMemoryRepository Tests
- Combine all related tests into a single class with nested classes for each method under test
